backend/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
paho-mqtt
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
pyyaml
websockets
psutil
